Capacity-aware migrations
Currently, the scheduler uses a global EDF scheduling algorithm,
migrating tasks to CPU cores without considering the core capacity
and the task utilization. This works well on homogeneous systems
(SCHED_DEADLINE tasks are guaranteed to have a bounded tardiness),
but presents some issues on heterogeneous systems. For example,
a SCHED_DEADLINE task might be migrated on a core that has not enough
processing capacity to correctly serve the task (think about a
task with runtime 70ms and period 100ms migrated to a core with
processing capacity 0.5)

This commit is a first step to address the issue: when a task wakes
up or migrates away from a CPU core, the scheduler tries to find an
idel core having enough processing capacity to serve the task.